apps/media-forensics/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
requests==2.31.0
pillow==10.3.0
imagehash==4.3.1
exifread==3.0.0
opencv-python-headless==4.10.0.84
numpy>=1.26,<2.0
